ai-utils/requirements.txt
anthropic>=0.65.0
python-dotenv>=1.0.0
